feat: invoke different personalities based on regex matching
