--- cogniq/bots/bing_search/__init__.py
@@ -1,1 +1,0 @@
-from .start import start
