--- cogniq/bots/chat_anthropic/__init__.py
@@ -1,1 +1,0 @@
-from .chat import start
