--- cogniq/bots/chat_gpt4/__init__.py
@@ -1,1 +1,0 @@
-from .chat_gpt4 import start
