+++ cogniq/personalities/bing_search/__init__.py
@@ -1,0 +1,2 @@
+from .ask import ask_task
+from .wake_pattern import wake_pattern
