+++ cogniq/personalities/bing_search/config.py
@@ -10,4 +10,5 @@ Config = {
     "OPENAI_MAX_TOKENS_RETRIEVAL": os.environ.get("OPENAI_MAX_TOKENS_RETRIEVAL", 700),
     "OPENAI_MAX_TOKENS_PROMPT": os.environ.get("OPENAI_MAX_TOKENS_PROMPT", 1000),
     "OPENAI_MAX_TOKENS_RESPONSE": os.environ.get("OPENAI_MAX_TOKENS_RESPONSE", 800),
+    "BING_SUBSCRIPTION_KEY": os.environ.get("BING_SUBSCRIPTION_KEY"),
 }
