+++ cogniq/personalities/bing_search/wake_pattern.py
@@ -1,0 +1,3 @@
+import re
+
+wake_pattern = re.compile(r"\b(search|bing|websearch|google)\b", re.IGNORECASE)
