+++ cogniq/personalities/chat_anthropic/wake_pattern.py
@@ -1,0 +1,3 @@
+import re
+
+wake_pattern = re.compile(r"\b(claude|anthropic)\b", re.IGNORECASE)
