+++ cogniq/personalities/chat_gpt4/wake_pattern.py
@@ -1,0 +1,3 @@
+import re
+
+wake_pattern = re.compile(r"\b(chatgpt|gpt|openai)\b", re.IGNORECASE)
